Revert "CSIT214-49 create homepage"
