--- data/data_countries.ts
@@ -1,7 +1,0 @@
-export const data_countries = [
-  "Nepan",
-  "Myanmar",
-  "Netherlands",
-  "United State",
-  "Germany",
-];
