--- data/data_experience.ts
@@ -1,42 +1,0 @@
-export const data_experience = [
-  {
-    img: "homePageImg/EntireHome1.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome2.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome3.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome4.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome5.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome6.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome7.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-  {
-    img: "homePageImg/EntireHome8.svg",
-    title: "EntireHome",
-    detail: "156,700 properties",
-  },
-];
