--- data/data_vacation.ts
@@ -1,20 +1,0 @@
-export const data_vaction = [
-  {
-    img: "homePageImg/Lakethun1.svg",
-    title: "Lake Thun, Switzerland",
-    review: 1.2,
-    rating: 4.4,
-  },
-  {
-    img: "homePageImg/Lakethun2.svg",
-    title: "Lake Thun, Switzerland",
-    review: 1.2,
-    rating: 4.4,
-  },
-  {
-    img: "homePageImg/Lakethun3.svg",
-    title: "Lake Thun, Switzerland",
-    review: 1.2,
-    rating: 4.4,
-  },
-];
