+++ package.json
@@ -5,37 +5,23 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
-    "@mui/icons-material": "^5.10.6",
-    "@mui/material": "^5.10.6",
-    "@mui/x-data-grid-premium": "^5.17.3",
-    "@mui/x-date-pickers": "^5.0.2",
-    "@mui/x-date-pickers-pro": "^5.0.2",
-    "@mui/x-license-pro": "^5.17.0",
-    "dayjs": "^1.11.5",
-    "moment": "^2.29.4",
-    "next": "^12.3.1",
+    "@mui/material": "^5.10.4",
+    "next": "12.2.5",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "yarn": "^1.22.19"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@types/node": "18.7.13",
     "@types/react": "18.0.17",
     "@types/react-dom": "18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.38.0",
-    "@typescript-eslint/parser": "^5.38.0",
-    "eslint": "^8.0.1",
+    "eslint": "8.23.0",
     "eslint-config-next": "12.2.5",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-n": "^15.0.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.31.8",
-    "typescript": "^4.8.3"
+    "typescript": "4.8.2"
   }
 }
