+++ pages/_app.tsx
@@ -1,12 +1,8 @@
-import "../styles/globals.css";
-import type { AppProps } from "next/app";
-import { LicenseInfo } from "@mui/x-license-pro";
+import '../styles/globals.css'
+import type { AppProps } from 'next/app'
 
-LicenseInfo.setLicenseKey(
-  "x0jTPl0USVkVZV0SsMjM1kDNyADM5cjM2ETPZJVSQhVRsIDN0YTM6IVREJ1T0b9586ef25c9853decfa7709eee27a1e"
-);
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return <Component {...pageProps} />
 }
 
-export default MyApp;
+export default MyApp
