+++ styles/globals.css
@@ -15,20 +15,12 @@ a {
   box-sizing: border-box;
 }
 
-fieldset {
-  border-width: 0 !important;
-}
-
-div:before {
-  border-bottom: 0 !important;
-}
-
 @media (prefers-color-scheme: dark) {
   html {
     color-scheme: dark;
   }
-
   body {
+    color: white;
     background: white;
   }
 }
